requirements.txt
pytest==6.2.3
allure-pytest==2.8.31
docker==5.0.0
